cpu_pool_policy: reworked configuration for flat configmap/dir.
Reworked configuration keeping in mind that a volume-mounted
config map has a flat directory structure. Instead of keeping
node configuration in subdirectories we now have to keep all
node configuration in a single directory in multiple files.

The node pool configuration is now a JSON file where either
a "size" or a "cpus" entry should be given, corrsponding to
the old @n and "explicit cpus" notation. cpus can be set to
"*" which has the same semantics as before, it will collect
all unclaimed CPUs to the pool.

The pool policy uses the newly added stub.ConfigPicker interface
to pick the configuration for itself. ConfigPicker is also used
to get notifications about configuration updates and reconfigure
the pools when this happens.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>